Move Interests to the common crate
